Improved camera capture in Android.
- Using camera2 API.
- The physical flash is now working.
- Use white screen when the flash device is not avalible.
- Take into account the pixel stride when reading the video frames.

+++ StandAlone/src/recording.cpp
@@ -999,7 +999,8 @@ bool RecordingPrivate::canAccessStorage()
 
         context.callObjectMethod("requestPermissions",
                                  "([Ljava/lang/String;I)V",
-                                 permissionsArray, Ak::id());
+                                 permissionsArray,
+                                 jint(Ak::id()));
         QElapsedTimer timer;
         timer.start();
         static const int timeout = 5000;
